+++ packages/patrol/android/src/main/kotlin/pl/leancode/patrol/PatrolServer.kt
@@ -9,14 +9,24 @@ import org.http4k.server.KtorCIO
 import org.http4k.server.asServer
 
 class PatrolServer {
+    private val defaultPort = 8081
+
     private val envPortKey = "PATROL_PORT"
-    private val port: Int
     private var server: Http4kServer? = null
     private var automatorServer: AutomatorServer? = null
 
-    init {
-        port = arguments.getString(envPortKey)?.toInt() ?: 8081
-    }
+    val port: Int
+        get() {
+            val portStr = arguments.getString(envPortKey)
+            if (portStr == null) {
+                Logger.i("$envPortKey is null, falling back to default ($defaultPort)")
+                return defaultPort
+            }
+            return portStr.toIntOrNull() ?: run {
+                Logger.i("$envPortKey is not a valid integer, falling back to default ($defaultPort)")
+                defaultPort
+            }
+        }
 
     private val arguments get() = InstrumentationRegistry.getArguments()