+++ packages/patrol_cli/lib/src/commands/test.dart
@@ -119,6 +119,7 @@ class TestCommand extends PatrolCommand {
     }
 
     if (devices.length > 1) {
+      // TODO: Throw an error when running on more than 1 device
       _logger.warn('''
 Running on multiple devices is deprecated and will be removed in the future.
 See https://github.com/leancodepl/patrol/issues/1316 to learn more.
@@ -148,6 +149,8 @@ See https://github.com/leancodepl/patrol/issues/1316 to learn more.
       'PATROL_IOS_APP_NAME': config.ios.appName,
       'INTEGRATION_TEST_SHOULD_REPORT_RESULTS_TO_NATIVE': 'false',
       'PATROL_TEST_LABEL_ENABLED': displayLabel.toString(),
+      'PATROL_TEST_SERVER_PORT': super.testServerPort.toString(),
+      'PATROL_APP_SERVER_PORT': super.appServerPort.toString(),
     }.withNullsRemoved();
 
     final dartDefines = {...customDartDefines, ...internalDartDefines};
@@ -183,6 +186,8 @@ See https://github.com/leancodepl/patrol/issues/1316 to learn more.
       scheme: buildMode.createScheme(iosFlavor),
       configuration: buildMode.createConfiguration(iosFlavor),
       simulator: !device.real,
+      appServerPort: super.appServerPort,
+      testServerPort: super.testServerPort,
     );
 
     final macosOpts = MacOSAppOptions(